代码优化

+++ app/Services/ServerMonitorService.php
@@ -2,12 +2,17 @@
 
 declare(strict_types=1);
 
+/**
+ * @Developer: ck
+ * @Email: ck@eqray.com
+ */
+
 namespace App\Services;
 
 class ServerMonitorService
 {
     /**
-     * 获取完整服务器信息
+     * 获取完整服务器信息.
      * @return array<array-key, mixed>
      */
     public function getServerInfo(): array
@@ -21,7 +26,7 @@ class ServerMonitorService
     }
 
     /**
-     * 获取完整监控信息（内存 + PHP环境 + 磁盘 + 缓存）
+     * 获取完整监控信息（内存 + PHP环境 + 磁盘 + 缓存）.
      * @return array<array-key, mixed>
      */
     public function getFullInfo(): array
@@ -32,18 +37,18 @@ class ServerMonitorService
     }
 
     /**
-     * 获取缓存信息（OPcache）
+     * 获取缓存信息（OPcache）.
      * @return array<array-key, mixed>
      */
     public function getCacheInfo(): array
     {
         $opcache = [];
         if (function_exists('opcache_get_status')) {
-            $status = opcache_get_status(false) ?: [];
-            $config = opcache_get_configuration()['directives'] ?? [];
-            $mem    = $status['memory_usage'] ?? [];
-            $used   = ($mem['used_memory'] ?? 0) + ($mem['wasted_memory'] ?? 0);
-            $free   = $mem['free_memory'] ?? 0;
+            $status  = opcache_get_status(false) ?: [];
+            $config  = opcache_get_configuration()['directives'] ?? [];
+            $mem     = $status['memory_usage']                   ?? [];
+            $used    = ($mem['used_memory'] ?? 0) + ($mem['wasted_memory'] ?? 0);
+            $free    = $mem['free_memory'] ?? 0;
             $opcache = [
                 'enabled'        => $status['opcache_enabled'] ?? false,
                 'memory_used'    => $this->formatBytes($used),
@@ -51,19 +56,17 @@ class ServerMonitorService
                 'memory_total'   => $this->formatBytes($used + $free),
                 'hit_rate'       => round($status['opcache_statistics']['opcache_hit_rate'] ?? 0, 2) . '%',
                 'cached_scripts' => $status['opcache_statistics']['num_cached_scripts'] ?? 0,
-                'max_files'      => $config['opcache.max_accelerated_files'] ?? 0,
+                'max_files'      => $config['opcache.max_accelerated_files']            ?? 0,
             ];
         }
         return ['opcache' => $opcache, 'php_version' => PHP_VERSION];
     }
 
     /**
-     * 获取 PHP 环境信息（供 MonitorController.php 使用）
-     *
-     * @return array
+     * 获取 PHP 环境信息（供 MonitorController.php 使用）.
      */
     /**
-     * 获取 PHP 信息
+     * 获取 PHP 信息.
      *
      * @return array<array-key, mixed>
      */
@@ -73,7 +76,7 @@ class ServerMonitorService
     }
 
     /**
-     * 清理缓存（OPcache + stat cache）
+     * 清理缓存（OPcache + stat cache）.
      *
      * @return array<array-key, mixed>
      */
@@ -91,7 +94,7 @@ class ServerMonitorService
 
     /**
      * 内存信息
-     * 兼容 Linux / Windows
+     * 兼容 Linux / Windows.
      *
      * @return array<array-key, mixed>
      */
@@ -103,13 +106,13 @@ class ServerMonitorService
         if (PHP_OS_FAMILY === 'Windows') {
             // 使用 PowerShell 获取内存信息（wmic 已在新版 Windows 中移除）
             $psCommand = 'powershell -Command "Get-CimInstance Win32_OperatingSystem | Select-Object TotalVisibleMemorySize, FreePhysicalMemory | ConvertTo-Json"';
-            $output = shell_exec($psCommand);
-            
+            $output    = shell_exec($psCommand);
+
             if ($output) {
                 $memData = json_decode($output, true);
                 if (is_array($memData)) {
-                    $totalBytes = (int)($memData['TotalVisibleMemorySize'] ?? 0) * 1024;
-                    $freeBytes  = (int)($memData['FreePhysicalMemory'] ?? 0) * 1024;
+                    $totalBytes = (int) ($memData['TotalVisibleMemorySize'] ?? 0) * 1024;
+                    $freeBytes  = (int) ($memData['FreePhysicalMemory'] ?? 0)     * 1024;
                 }
             }
         } else {
@@ -118,8 +121,8 @@ class ServerMonitorService
                 $meminfo = file_get_contents('/proc/meminfo');
                 preg_match('/MemTotal:\s+(\d+)\s+kB/i', $meminfo, $total);
                 preg_match('/MemAvailable:\s+(\d+)\s+kB/i', $meminfo, $free);
-                $totalBytes = isset($total[1]) ? (int)$total[1] * 1024 : 0;
-                $freeBytes  = isset($free[1])  ? (int)$free[1]  * 1024 : 0;
+                $totalBytes = isset($total[1]) ? (int) $total[1] * 1024 : 0;
+                $freeBytes  = isset($free[1]) ? (int) $free[1]   * 1024 : 0;
             }
         }
 
@@ -132,12 +135,12 @@ class ServerMonitorService
             'used'  => $this->formatBytes($usedBytes),
             'free'  => $this->formatBytes($freeBytes),
             'php'   => $this->formatBytes($phpBytes),
-            'rate'  => (string)$rate,
+            'rate'  => (string) $rate,
         ];
     }
 
     /**
-     * PHP 及环境信息
+     * PHP 及环境信息.
      *
      * @return array<string, mixed>
      */
@@ -152,7 +155,7 @@ class ServerMonitorService
             'project_path'       => defined('BASE_PATH') ? BASE_PATH : getcwd(),
             'memory_limit'       => ini_get('memory_limit'),
             'max_execution_time' => ini_get('max_execution_time'),
-            'error_reporting'    => $this->errorReportingToString((int)ini_get('error_reporting')),
+            'error_reporting'    => $this->errorReportingToString((int) ini_get('error_reporting')),
             'display_errors'     => ini_get('display_errors') ? 'On' : 'Off',
             'upload_max_filesize'=> ini_get('upload_max_filesize'),
             'post_max_size'      => ini_get('post_max_size'),
@@ -163,7 +166,7 @@ class ServerMonitorService
 
     /**
      * 磁盘分区信息
-     * 兼容 Linux / Windows
+     * 兼容 Linux / Windows.
      *
      * @return array<array-key, mixed>
      */
@@ -175,7 +178,7 @@ class ServerMonitorService
             // Windows: 枚举盘符
             foreach (range('A', 'Z') as $letter) {
                 $path = $letter . ':\\';
-                if (!is_dir($path)) {
+                if (! is_dir($path)) {
                     continue;
                 }
                 $total     = disk_total_space($path);
@@ -214,10 +217,10 @@ class ServerMonitorService
                 }
             } else {
                 // fallback: 只读根目录
-                $total = disk_total_space('/');
-                $free  = disk_free_space('/');
-                $used  = $total - $free;
-                $pct   = $total > 0 ? round($used / $total * 100, 1) : 0;
+                $total   = disk_total_space('/');
+                $free    = disk_free_space('/');
+                $used    = $total - $free;
+                $pct     = $total > 0 ? round($used / $total * 100, 1) : 0;
                 $disks[] = [
                     'filesystem'     => 'rootfs',
                     'size'           => $this->formatBytes($total),
@@ -232,10 +235,9 @@ class ServerMonitorService
         return $disks;
     }
 
-
     /**
      * 获取 CPU 信息
-     * 兼容 Linux / Windows
+     * 兼容 Linux / Windows.
      *
      * @return array<string, mixed>
      */
@@ -243,59 +245,58 @@ class ServerMonitorService
     {
         $info = [
             'architecture'     => 'unknown',
-            'model'           => 'unknown',
-            'physical_cores'  => 0,
-            'logical_cores'   => 0,
-            'max_frequency'   => 'unknown',
-            'current_usage'   => 0,
+            'model'            => 'unknown',
+            'physical_cores'   => 0,
+            'logical_cores'    => 0,
+            'max_frequency'    => 'unknown',
+            'current_usage'    => 0,
         ];
 
         if (PHP_OS_FAMILY === 'Windows') {
             // Windows: 使用 PowerShell 获取 CPU 信息（wmic 已在新版 Windows 中移除）
             $psCommand = 'powershell -Command "Get-CimInstance Win32_Processor | Select-Object Name, NumberOfCores, NumberOfLogicalProcessors, MaxClockSpeed, Architecture | ConvertTo-Json"';
-            $output = shell_exec($psCommand);
-            
+            $output    = shell_exec($psCommand);
+
             if ($output) {
                 $cpuData = json_decode($output, true);
                 if (is_array($cpuData) && isset($cpuData['Name'])) {
                     // 单个 CPU
                     $cpuData = [$cpuData];
                 }
-                if (!empty($cpuData)) {
-                    $cpu = is_array($cpuData[0]) ? $cpuData[0] : $cpuData;
-                    $info['model'] = $cpu['Name'] ?? 'unknown';
-                    $info['physical_cores'] = (int)($cpu['NumberOfCores'] ?? 0);
-                    $info['logical_cores'] = (int)($cpu['NumberOfLogicalProcessors'] ?? 0);
-                    $maxMhz = $cpu['MaxClockSpeed'] ?? 0;
-                    $info['max_frequency'] = $maxMhz > 0 ? round($maxMhz / 1000, 2) . ' GHz' : 'unknown';
-                    
+                if (! empty($cpuData)) {
+                    $cpu                    = is_array($cpuData[0]) ? $cpuData[0] : $cpuData;
+                    $info['model']          = $cpu['Name'] ?? 'unknown';
+                    $info['physical_cores'] = (int) ($cpu['NumberOfCores'] ?? 0);
+                    $info['logical_cores']  = (int) ($cpu['NumberOfLogicalProcessors'] ?? 0);
+                    $maxMhz                 = $cpu['MaxClockSpeed'] ?? 0;
+                    $info['max_frequency']  = $maxMhz > 0 ? round($maxMhz / 1000, 2) . ' GHz' : 'unknown';
+
                     // 架构映射
                     $archMap = [
-                        0 => 'x86',
-                        1 => 'MIPS',
-                        2 => 'Alpha',
-                        5 => 'ARM',
-                        6 => 'IA64',
-                        9 => 'x64',
+                        0  => 'x86',
+                        1  => 'MIPS',
+                        2  => 'Alpha',
+                        5  => 'ARM',
+                        6  => 'IA64',
+                        9  => 'x64',
                         12 => 'ARM64',
                     ];
-                    $info['architecture'] = $archMap[(int)($cpu['Architecture'] ?? 9)] ?? 'unknown';
+                    $info['architecture'] = $archMap[(int) ($cpu['Architecture'] ?? 9)] ?? 'unknown';
                 }
             }
-            
+
             // 获取 CPU 利用率
             $info['current_usage'] = $this->getWindowsCpuUsage();
-
         } else {
             // Linux: 读取 /proc/cpuinfo
             if (is_readable('/proc/cpuinfo')) {
                 $cpuinfo = file_get_contents('/proc/cpuinfo');
-                
+
                 // 获取 CPU 型号
                 if (preg_match('/model name\s*:\s*(.+)/i', $cpuinfo, $matches)) {
                     $info['model'] = trim($matches[1]);
                 }
-                
+
                 // 获取架构
                 if (preg_match('/architecture\s*:\s*(.+)/i', $cpuinfo, $matches)) {
                     $info['architecture'] = trim($matches[1]);
@@ -304,35 +305,35 @@ class ServerMonitorService
                 } else {
                     $info['architecture'] = trim(shell_exec('uname -m') ?? 'unknown');
                 }
-                
+
                 // 统计物理核心数和逻辑核心数
                 $physicalIds = [];
-                $coreIds = [];
-                $processors = 0;
-                
+                $coreIds     = [];
+                $processors  = 0;
+
                 foreach (explode("\n", $cpuinfo) as $line) {
                     if (preg_match('/^processor\s*:\s*(\d+)/i', $line, $m)) {
-                        $processors = max($processors, (int)$m[1] + 1);
+                        $processors = max($processors, (int) $m[1] + 1);
                     }
                     if (preg_match('/^physical id\s*:\s*(\d+)/i', $line, $m)) {
-                        $physicalIds[(int)$m[1]] = true;
+                        $physicalIds[(int) $m[1]] = true;
                     }
                     if (preg_match('/^core id\s*:\s*(\d+)/i', $line, $m)) {
-                        $coreIds[(int)$m[1]] = true;
+                        $coreIds[(int) $m[1]] = true;
                     }
                 }
-                
-                $physicalCount = count($physicalIds);
+
+                $physicalCount          = count($physicalIds);
                 $info['physical_cores'] = $physicalCount > 0 ? $physicalCount * count($coreIds) : $processors;
-                $info['logical_cores'] = $processors;
-                
+                $info['logical_cores']  = $processors;
+
                 // 获取 CPU 最大频率
                 if (preg_match('/cpu MHz\s*:\s*([\d.]+)/i', $cpuinfo, $matches)) {
-                    $info['max_frequency'] = round((float)$matches[1] / 1000, 2) . ' GHz';
+                    $info['max_frequency'] = round((float) $matches[1] / 1000, 2) . ' GHz';
                 } elseif (preg_match('/cpu max MHz\s*:\s*([\d.]+)/i', $cpuinfo, $matches)) {
-                    $info['max_frequency'] = round((float)$matches[1] / 1000, 2) . ' GHz';
+                    $info['max_frequency'] = round((float) $matches[1] / 1000, 2) . ' GHz';
                 }
-                
+
                 // 获取 CPU 利用率
                 $info['current_usage'] = $this->getLinuxCpuUsage();
             }
@@ -342,29 +343,27 @@ class ServerMonitorService
     }
 
     /**
-     * 获取 Windows CPU 利用率
+     * 获取 Windows CPU 利用率.
      */
     private function getWindowsCpuUsage(): float
     {
         // 使用 PowerShell 获取 CPU 利用率
         $psCommand = 'powershell -Command "(Get-CimInstance Win32_PerfFormattedData_PerfOS_Processor | Where-Object { $_.Name -eq \'_Total\' }).PercentProcessorTime"';
-        $output = shell_exec($psCommand);
-        
+        $output    = shell_exec($psCommand);
+
         if ($output) {
             $usage = trim($output);
             if (is_numeric($usage)) {
-                return round((float)$usage, 1);
+                return round((float) $usage, 1);
             }
         }
-        
+
         return 0;
     }
 
     /**
-     * 获取 Linux CPU 利用率
+     * 获取 Linux CPU 利用率.
      */
-            /**
-             */
     private function getLinuxCpuUsage(): float
     {
         // 方法1: 读取 /proc/stat
@@ -372,29 +371,29 @@ class ServerMonitorService
             $stat1 = $this->readCpuStats();
             usleep(100000); // 等待 100ms
             $stat2 = $this->readCpuStats();
-            
+
             $totalDiff = $stat2['total'] - $stat1['total'];
-            $idleDiff = $stat2['idle'] - $stat1['idle'];
-            
+            $idleDiff  = $stat2['idle']  - $stat1['idle'];
+
             if ($totalDiff > 0) {
                 return round(($totalDiff - $idleDiff) / $totalDiff * 100, 1);
             }
         }
-        
+
         // 方法2: 使用 top 命令
         $output = shell_exec("top -bn1 2>/dev/null | grep 'Cpu(s)' | awk '{print \$2}'");
         if ($output) {
             $usage = trim($output);
             if (is_numeric($usage)) {
-                return round((float)$usage, 1);
+                return round((float) $usage, 1);
             }
         }
-        
+
         return 0;
     }
 
     /**
-     * 读取 /proc/stat 中的 CPU 统计数据
+     * 读取 /proc/stat 中的 CPU 统计数据.
      *
      * @return array<string, mixed>
      */
@@ -402,35 +401,35 @@ class ServerMonitorService
     {
         $stat = file_get_contents('/proc/stat');
         if (preg_match('/cpu\s+(\d+)\s+(\d+)\s+(\d+)\s+(\d+)\s+(\d+)\s+(\d+)\s+(\d+)/', $stat, $matches)) {
-            $user = (int)$matches[1];
-            $nice = (int)$matches[2];
-            $system = (int)$matches[3];
-            $idle = (int)$matches[4];
-            $iowait = (int)$matches[5];
-            $irq = (int)$matches[6];
-            $softirq = (int)$matches[7];
-            
+            $user    = (int) $matches[1];
+            $nice    = (int) $matches[2];
+            $system  = (int) $matches[3];
+            $idle    = (int) $matches[4];
+            $iowait  = (int) $matches[5];
+            $irq     = (int) $matches[6];
+            $softirq = (int) $matches[7];
+
             $total = $user + $nice + $system + $idle + $iowait + $irq + $softirq;
-            
+
             return [
                 'total' => $total,
                 'idle'  => $idle,
             ];
         }
-        
+
         return ['total' => 0, 'idle' => 0];
     }
 
     // ==================== helpers ====================
 
-    private function formatBytes(int|float $bytes): string
+    private function formatBytes(float|int $bytes): string
     {
         if ($bytes <= 0) {
             return '0 B';
         }
         $units = ['B', 'KB', 'MB', 'GB', 'TB'];
-        $i = (int)floor(log($bytes, 1024));
-        $i = min($i, count($units) - 1);
+        $i     = (int) floor(log($bytes, 1024));
+        $i     = min($i, count($units) - 1);
         return round($bytes / (1024 ** $i), 2) . ' ' . $units[$i];
     }
 
@@ -451,6 +450,6 @@ class ServerMonitorService
                 return "{$name} ({$level})";
             }
         }
-        return (string)$level;
+        return (string) $level;
     }
 }
